Pull group ctx createChannel into `shared`
