+++ packages/shared/src/logic/index.ts
@@ -7,3 +7,4 @@ export * from './activity';
 export * from './branch';
 export * from './deeplinks';
 export * from './analytics';
+export * from './useCreateChannel';
